chore: fix lint findings; raise app-size budget to measured 16.3 MB
The raw-response consumer moves to AnyTaskCancellable behind an
armRawResponses helper (no_manual_task_management, connect complexity);
DataTrackPublishOptions adopts @objcMembers; the publish-focused tests
split into DataTrackPublishTests (file/type length). Budget follows the
measured 16.28 MB delta after the data-track and nanopb changes.
