+++ Tests/LiveKitCoreTests/DataTrack/DataTrackTests.swift
@@ -83,103 +83,6 @@ struct DataTrackTests {
         }
     }
 
-    // MARK: - Frame Metadata
-
-    /// Publishing with declared frame metadata (schema + encoding) succeeds and frames flow.
-    /// The metadata is sent to the SFU (parity with rust-sdks `DataTrackOptions`), but servers
-    /// don't echo it back in track info yet (as of livekit-server 1.13.5), so the round-trip via
-    /// ``DataTrackInfo/schema``/``DataTrackInfo/frameEncoding`` isn't asserted here.
-    @Test
-    func publishWithFrameMetadata() async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(canPublishData: true),
-            RoomTestingOptions(canSubscribe: true),
-        ]) { rooms in
-            let watcher = DataTrackWatcher(expectedName: "typed")
-            rooms[1].delegates.add(delegate: watcher)
-
-            let schema = DataTrackSchemaId(name: "my-schema", encoding: .jsonSchema)
-            let options = DataTrackPublishOptions(schema: schema, frameEncoding: .json)
-            let track = try await rooms[0].localParticipant.publishDataTrack(name: "typed", options: options)
-            #expect(track.isPublished)
-
-            let remoteTrack = try await watcher.waitForTrack()
-            let stream = try await remoteTrack.subscribe()
-            let payload = Data("{}".utf8)
-            try track.tryPush(frame: DataTrackFrame(payload: payload))
-            let frame = try #require(await stream.next())
-            #expect(frame.payload == payload)
-        }
-    }
-
-    // MARK: - Without E2EE
-
-    /// Data tracks work without E2EE: the publication is not marked encrypted and frames arrive
-    /// as sent. Pins the no-encryption path, which every other test here skips — `withRooms`
-    /// enables E2EE by default.
-    @Test
-    func publishAndReceiveWithoutE2ee() async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(isE2eeEnabled: false, canPublishData: true),
-            RoomTestingOptions(isE2eeEnabled: false, canSubscribe: true),
-        ]) { rooms in
-            let watcher = DataTrackWatcher(expectedName: "plaintext")
-            rooms[1].delegates.add(delegate: watcher)
-
-            let track = try await rooms[0].localParticipant.publishDataTrack(name: "plaintext")
-            let remoteTrack = try await watcher.waitForTrack()
-            #expect(!remoteTrack.info.usesE2ee)
-
-            let stream = try await remoteTrack.subscribe()
-            let payload = Data([0x0B, 0x0E])
-            try track.tryPush(frame: DataTrackFrame(payload: payload))
-            let frame = try #require(await stream.next())
-            #expect(frame.payload == payload)
-        }
-    }
-
-    // MARK: - Publish Duplicate Name
-
-    @Test
-    func publishDuplicateName() async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(canPublishData: true),
-        ]) { rooms in
-            let room = rooms[0]
-            let first = try await room.localParticipant.publishDataTrack(name: "dup")
-            #expect(first.isPublished)
-            await #expect(throws: DataTrackPublishError.self) {
-                _ = try await room.localParticipant.publishDataTrack(name: "dup")
-            }
-            _ = first.isPublished // keep "dup" published while the duplicate attempt runs
-        }
-    }
-
-    // MARK: - Publish Unauthorized
-
-    @Test
-    func publishUnauthorized() async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(canPublishData: false),
-        ]) { rooms in
-            let room = rooms[0]
-            await #expect(throws: DataTrackPublishError.self) {
-                _ = try await room.localParticipant.publishDataTrack(name: "unauth")
-            }
-        }
-    }
-
-    // MARK: - Publish While Disconnected
-
-    /// Publishing on a room that was never connected fails with the data-track error type
-    /// (`.disconnected`), matching Rust — not a generic `LiveKitError`.
-    @Test
-    func publishWhileDisconnectedThrows() async throws {
-        await #expect(throws: DataTrackPublishError.self) {
-            _ = try await Room().localParticipant.publishDataTrack(name: "nope")
-        }
-    }
-
     // MARK: - Published State
 
     @Test
